Refactor overlay controls into separate module
